extract generation stop logic into dedicated microcrate
